Reduce object size by switching from deque -> vector.
sizeof(deque) = 48, sizeof(vector) = 24 (when empty).
